+++ tests/regression/io_addr/common.h
@@ -5,8 +5,8 @@
 
 typedef struct {
   uint32_t num_points;
-  uint32_t src_addr;
-  uint32_t dst_addr;  
+  uint64_t src_addr;
+  uint64_t dst_addr;  
 } kernel_arg_t;
 
 #endif
