+++ tests/regression/io_addr/kernel.cpp
@@ -4,7 +4,7 @@
 #include "common.h"
 
 void kernel_body(int task_id, kernel_arg_t* __UNIFORM__ arg) {
-	uint32_t* src_ptr = (uint32_t*)arg->src_addr;
+	uint64_t* src_ptr = (uint64_t*)arg->src_addr;
 	uint32_t* dst_ptr = (uint32_t*)arg->dst_addr;
 
 	int32_t* addr_ptr = (int32_t*)(src_ptr[task_id]);
